refactor(responses): group pending snapshot tracking and drop a redundant chaining test
